examples/computer_use_rl/requirements.txt
# Minimal deps for the VNC/KVM computer-use RL demo
vncdotool>=1.2.0
Pillow>=9.0.0
numpy>=1.23.0
